Renamed team_effect_apply, team_viewer, and team_serialize to effects, viewer, and serialize respectively.
