+++ src/lib/teams/mod.rs
@@ -1,10 +1,10 @@
 //! [`Team`](crate::teams::team::Team) battle and effect logic..
 
+/// [`Effect`](crate::Effect) application to one or more [`Team`](crate::Team)s.
+pub mod effects;
+/// Serialize a [`Team`](crate::Team) using [`serde_json`].
+pub mod serialize;
 /// [`Team`](crate::Team) of [`Pet`](crate::Pet)s.
 pub mod team;
-/// [`Effect`](crate::Effect) application to one or more [`Team`](crate::Team)s.
-pub mod team_effect_apply;
-/// Serialize a [`Team`](crate::Team) using [`serde_json`].
-pub mod team_serialize;
 /// View a [`Team`](crate::Team)'s [`Pet`](crate::Pet)s.
-pub mod team_viewer;
+pub mod viewer;
