+++ src/lib/tests/test_food.rs
@@ -10,7 +10,7 @@ use crate::{
     pets::{combat::PetCombat, names::PetName, pet::Pet},
     teams::{
         team::{Team, TeamFightOutcome},
-        team_viewer::TeamViewer,
+        viewer::TeamViewer,
     },
     tests::common::test_ant_team,
     Effect,
