+++ src/lib/tests/test_team_t2.rs
@@ -1,7 +1,7 @@
 use crate::{
     effects::{state::Position, trigger::TRIGGER_START_BATTLE},
     pets::names::PetName,
-    teams::{team::TeamFightOutcome, team_viewer::TeamViewer},
+    teams::{team::TeamFightOutcome, viewer::TeamViewer},
     tests::common::{
         test_ant_team, test_atlantic_puffin_team, test_bat_team, test_crab_team, test_dodo_team,
         test_dove_team, test_elephant_peacock_team, test_flamingo_team, test_hedgehog_team,
