+++ src/lib/tests/test_team_t3.rs
@@ -2,7 +2,7 @@ use crate::{
     effects::{state::Status, stats::Statistics, trigger::TRIGGER_START_BATTLE},
     foods::{food::Food, names::FoodName},
     pets::names::PetName,
-    teams::{team::TeamFightOutcome, team_viewer::TeamViewer},
+    teams::{team::TeamFightOutcome, viewer::TeamViewer},
     tests::common::{
         test_aardvark_team, test_ant_team, test_badger_team, test_bear_team, test_blobfish_team,
         test_blowfish_rally_team, test_blowfish_team, test_camel_team, test_clownfish_team,
